Prepare FilterOptions to enable autofocus
